backend/requirements.txt
fastapi[standard]
langchain-text-splitters
langchain_pinecone
langchain_openai
langchain
pandas
cohere
pdfminer.six
langchain_experimental